Add Look up table. Good settings for upwards wall jump

+++ lib/experiment.h
@@ -2,4 +2,4 @@ void expStart();
 void expFlow();
 char footContact(void);
 long calibPos(char idx);
-long crankFromFemur(void);
+unsigned int crankFromFemur(void);
